feat(web): make push-replica caching policy explicit with one root event-sync module (resolves #321)
